chore: build runtime
