+++ apps/electron-client/package.json
@@ -27,7 +27,7 @@
     "@radix-ui/react-popover": "^1.0.2",
     "@raycast/api": "^1.45.2",
     "@svgr/webpack": "^6.5.1",
-    "@types/react": "^18.0.26",
+    "@types/react": "18.0.9",
     "@types/react-dom": "^18.0.10",
     "@typescript-eslint/eslint-plugin": "^5.47.0",
     "@typescript-eslint/parser": "^5.47.0",
@@ -37,6 +37,7 @@
     "css-loader": "^6.7.3",
     "electron": "22.0.0",
     "electron-devtools-installer": "^3.2.0",
+    "eslint": "^8.30.0",
     "fork-ts-checker-webpack-plugin": "^7.2.14",
     "highlight.js": "^11.7.0",
     "node-loader": "^2.0.0",
@@ -45,12 +46,14 @@
     "postcss-preset-env": "^7.8.3",
     "postcss-scss": "^4.0.6",
     "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rpc-websockets": "^7.5.0",
     "sass": "^1.57.1",
     "sass-loader": "^13.2.0",
     "style-loader": "^3.3.1",
     "tailwindcss": "^3.2.4",
     "ts-loader": "^9.4.2",
+    "typescript": "^4.9.4",
     "webpack": "5",
     "zustand": "^4.1.5"
   },
