--- packages/blast-runtime/index.ts
@@ -1,10 +1,0 @@
-// import Command from "../../../../examples/todo-list/src/index";
-import { run } from "../renderer/run";
-import { runServer } from "../server";
-import { connect } from "../utils/connectDevtools";
-
-connect();
-
-run(Command, {
-  server: runServer(),
-});
