+++ packages/blast-runtime/src/@types/raycast.d.ts
@@ -1,4 +1,4 @@
-// Create a module alias for raycast-original to @raycast/api
-declare module "@raycast/api" {
-  export * from "raycast-original";
-}
+// // Create a module alias for raycast-original to @raycast/api
+// declare module "@raycast/api" {
+//   export * from "raycast-original";
+// }
