+++ packages/blast-runtime/src/index.ts
@@ -1,0 +1,11 @@
+import { runServer } from "blast-renderer";
+
+import { App } from "./App";
+import { connect } from "./utils/connectDevtools";
+import { run } from "./utils/run";
+
+connect();
+
+run(App, {
+  server: runServer(),
+});
