+++ packages/blast-runtime/src/utils/window.js
